btrfs: support device name lookup in forget
btrfs forget assumes the device still exists in the block layer and
that we can lookup its dev_t. For handling some tricky cases with
changing devt across device recreation, we need udev rules that run on
device removal. However, at that point, there is no node to lookup, so
we need to rely on the cached name. Refactor the forget code to handle
this case, while still preferring to use dev_t when possible.

Tested by the new fstest: XXX (assumes new btrfs-progs which takes
advantage of this from device scan -u and udev.

Signed-off-by: Boris Burkov <boris@bur.io>