btrfs: add parents and children to qgroup sysfs
Adding this information lets users more or less completely inspect the
state of qgroups and replicate "btrfs qgroup list" via sysfs. If we have
the kobjects, might as well flesh them out to the utmost!

I experimented with adding a public helper function to use the qgroup
iterator to get the parents/children, but that had some weird semantics
with locking and calling clean. With that in mind, I ultimately decided
that putting btrfs_qgroup_list in the header was simpler and more closely
matched the intended use case.

Signed-off-by: Boris Burkov <boris@bur.io>