btrfs: hold merge refcount during extent_map merge
In debugging some corrupt squashfs files, we observed symptoms of
corrupt page cache pages but correct on-disk contents. Further
investigation revealed that the exact symptom was a correct page
followed by an incorrect, duplicate, page. This got us thinking about
extent maps.

commit ac05ca913e9f ("Btrfs: fix race between using extent maps and merging them")
enforces a reference count on the primary `em` extent_map being merged,
as that one gets modified.

However, since,
commit 3d2ac9922465 ("btrfs: introduce new members for extent_map")
both 'em' and 'merge' get modified, which introduced the same race and
need for refcounting for 'merge'.

We were able to reproduce this by looping the affected squashfs workload
in parallel on a bunch of separate btrfs-es while also dropping caches.

The fix is two-fold:
- check the refs > 2 criterion for both 'em' and 'merge'
- actually refcount 'merge' which we currently grab without a refcount
  via rb_prev or rb_next.

The other option we considered, but did not implement, was to stop
modifying 'merge' as it gets deleted immediately after. However, it was
less clear what effect that would have on the racing thread holding the
reference, so we went with this fix.

Fixes: 3d2ac9922465 ("btrfs: introduce new members for extent_map")
Signed-off-by: Omar Sandoval <osandov@fb.com>
Signed-off-by: Boris Burkov <boris@bur.io>