mm/filemap: global memcg vmstats
The new filemap_add_folio_nocharge() function produces pages which are
attached to a mapping but aren't accounted for by any cgroup. Therefore,
it is valuable to track them with a new vmstat. Similarly, pages in the
root cgroup are somewhat difficult to track, as the root cgroup
recursively rolls up all the page cache from the subgroups, so tracking
the root file pages separately can be helpful for assessing system page
cache usage.

Suggested-by: Shakeel Butt <shakeel.butt@linux.dev>
Signed-off-by: Boris Burkov <boris@bur.io>